Use correct class for data serialized as Serialized[Distribution]
We use Serialized[Distribution] types to avoid saving unnecessary
(and incorrect) RNG state, and each Distribution subclass may also
subclass Distribution.Data. Therefore we need to ensure that when
deserializing to e.g. Serialized[MvDistribution] (which is only
implemented via subclasses), that it always redirects to the correct
subclass (like Serialized[MvNormalDistribution]). This is what this
commit does, by adding a method `valid_distname` to each subclass of
Distribution.Data, and some __new__ magic to check call this method
for all subclasses and find the subclass which matches the data being
deserialized.