Rename 'json_encoder' -> 'reduce'
Even though it aligns with usage in Pydantic, I find that 'json_encoder'
too strongly suggests a function which returns a JSON string.
'reduce' better suggests a function which you may apply multiple times,
every time reducing the object to a set of simpler types. This is
analogous to the __reduce__ method used by pickle.